+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0] - June 20 2019.
 
-* TODO: Describe initial release.
+* Add null checks to all parameters.
+* Updated documentation.
+
+### New Features
+
+* Add *splashColor* parameter to define custom splash colors.
+* Add *minRadius* and *maxRadius* parameters to define splash size constraints.